+++ server/src/main/java/com/pyamsoft/tetherfi/server/proxy/session/tcp/TransportProtocols.kt
@@ -81,7 +81,7 @@ private suspend fun enforceBandwidthLimit(
     client: TetherClient,
     bandwidthLimit: Long,
     startTimeNanos: Long,
-    read: Int
+    read: Long
 ): Long {
   val bytesCopied = client.limiter.updateAndGet(read)
